Extract toolbar management from ApplicationHelper
We don't need to define all these globally-visible helper methods -- the
usable surface is actually two very narrow methods. Their respective
supporting methods don't even overlap.

This is a straightforward relocation -- other than adding some trivial
plumbing, the code is all moving untouched.
